refactor: :recycle: Improve Endpoints in Aspire Dashboard as per 9.3
1. Replaced existing Endpoint URL for Event Management service with Scalar OpenAPI | 2. Moved Clarity Config under services

Fixes #143
