+++ Visage.FrontEnd/Visage.FrontEnd.Web/Services/ClarityConfig.cs
@@ -1,4 +1,4 @@
-namespace Visage.FrontEnd.Web;
+namespace Visage.FrontEnd.Web.Services;
 
 public class ClarityConfig
 {
